requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django
django-ckeditor==6.6.1
django-crispy-forms==2.0
django-filter==23.2
django-js-asset==2.1.0
gunicorn==20.1.0
Pillow==9.5.0
qrcode[pil]
pandas 
openpyxl
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
whitenoise==6.5.0